Resources/python/dashboard/requirements.txt
# Server
Flask==3.1.0

# Configuration
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.0

# RDF and Triple Store
rdflib==7.1.1
requests==2.32.3
pytz~=2024.2
setuptools~=68.2.0